Add: autoPlayのオプション化 Fix: 初回起動時にオプションが設定されていない不具合
