+++ Assets/Scripts/AutoControllerFlag.cs
@@ -1,0 +1,8 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class AutoControllerFlag : MonoBehaviour
+{
+    public static bool doAutoPlay = false;
+}
