+++ Assets/Scripts/AutoControllerFlag.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4bc4c55e59f602746883bbd549ec5760
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
