+++ Assets/Scripts/GlobalConfigInitializer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 204b955f3e3ab5140971fbbae232ee63
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
