+++ Assets/Scripts/GlobalConfigSetter.cs
@@ -5,17 +5,6 @@ using UnityEngine.UI;
 
 public class GlobalConfigSetter : MonoBehaviour
 {
-    // Start is called before the first frame update
-    void Start()
-    {
-        if (!PlayerPrefs.HasKey("useUI"))
-            PlayerPrefs.SetInt("useUI", 1);
-        if (!PlayerPrefs.HasKey("useCentralIcon"))
-            PlayerPrefs.SetInt("useCentralIcon", 1);
-
-        PlayerPrefs.Save();
-    }
-
     public void SetTogglePrefs(Toggle t)
     {
         if (t.isOn)
@@ -24,5 +13,4 @@ public class GlobalConfigSetter : MonoBehaviour
             PlayerPrefs.SetInt(t.gameObject.name, 0);
         PlayerPrefs.Save();
     }
-
 }
